feat(make admin ui): refactor code
