+++ src/shared/components/exercise-content/ExerciseContent.tsx
@@ -1,6 +1,6 @@
 import { ExerciseType, useDictionaryStore, EXERCISE_FORMATE } from '@/store/dictionary';
 import { Typography } from '@/ui-components/Typography';
-import { classes } from '@/shared/utils/classes';
+import { classes } from '@/shared/services/classes';
 import { Button } from '@/ui-components/Button';
 import { speak } from '@/shared/utils/speak';
 import { Icon } from '@/ui-components/Icon';
