+++ src/ui-components/Menu/Menu.tsx
@@ -3,7 +3,7 @@ import { useState, ReactNode } from 'react';
 import type { IconVariantsType } from '../Icon/Icon';
 import { Icon } from '@/ui-components/Icon';
 import { CircleButton } from '@/ui-components/CircleButton';
-import { classes } from '@/shared/utils/classes';
+import { classes } from '@/shared/services/classes';
 import styles from './menu.module.css';
 
 const Side = {
